🔨🤖 Move the Algolia facet-filter helpers into @ourworldindata/utils
Pure move, no behavior change: the country/topic/Featured-Metric facet-filter
builders leave site/search/searchUtils.tsx for
packages/@ourworldindata/utils/src/search/searchFacetFilters.ts, so the
public /api/search Cloudflare function — which can't import from site/ — can
share them in a later commit.

The only edits to the moved code are explicit return types, which
packages/@ourworldindata/* enforces and site/** doesn't. searchUtils.tsx
re-exports every moved name, so no call site outside these two files changes.
