+++ site/search/searchUtils.tsx
@@ -15,7 +15,6 @@ import {
     ScoredFilter,
     SearchResultType,
     SearchTopicType,
-    SearchFacetFilters,
     ChartRecordType,
     SearchChartHit,
     SearchUrlParam,
@@ -33,6 +32,14 @@ import {
     timeBoundToTimeBoundString,
     queryParamsToStr,
     omitUndefinedValues,
+    type SearchFacetAttribute,
+    getFilterNamesOfType,
+    setToFacetFilters,
+    formatDisjunctiveFacetFilters,
+    formatConjunctiveFacetFilters,
+    formatFeaturedMetricFacetFilter,
+    formatCountryFacetFilters,
+    formatTopicFacetFilters,
 } from "@ourworldindata/utils"
 import {
     generateSelectedEntityNamesParam,
@@ -377,20 +384,19 @@ export const DATA_CATALOG_ATTRIBUTES = [
     "explorerType",
 ]
 
-type SearchFacetAttribute =
-    | "tags" // also used on /latest
-    | "latestType" // used on /latest only
-    | "availableEntities"
-    | "datasetProducts"
-    | "datasetNamespaces"
-    | "datasetVersions"
-    | "datasetProducers"
-
-export function setToFacetFilters(
-    facetSet: Set<string>,
-    attribute: SearchFacetAttribute
-) {
-    return Array.from(facetSet).map((facet) => `${attribute}:${facet}`)
+// Re-exported for existing importers; the actual definitions live in
+// @ourworldindata/utils so functions/api/search/searchApi.ts (which can't
+// import from site/) shares them too, keeping Algolia facetFilters identical
+// between the site's search UI and the public /api/search endpoint.
+export type { SearchFacetAttribute }
+export {
+    getFilterNamesOfType,
+    setToFacetFilters,
+    formatDisjunctiveFacetFilters,
+    formatConjunctiveFacetFilters,
+    formatFeaturedMetricFacetFilter,
+    formatCountryFacetFilters,
+    formatTopicFacetFilters,
 }
 
 export function getSelectableTopics(
@@ -407,77 +413,12 @@ export function getSelectableTopics(
     return new Set()
 }
 
-export const formatDisjunctiveFacetFilters = (
-    facets: Set<string>,
-    attribute: SearchFacetAttribute
-): SearchFacetFilters => {
-    // disjunction mode (A OR B): [[attribute:"A", attribute:"B"]]
-    return [setToFacetFilters(facets, attribute)]
-}
-
-export const formatConjunctiveFacetFilters = (
-    facets: Set<string>,
-    attribute: SearchFacetAttribute
-): SearchFacetFilters => {
-    // conjunction mode (A AND B): [attribute:"A", attribute:"B"]
-    return setToFacetFilters(facets, attribute)
-}
-
-/**
- * Returns a facet filter that excludes Featured Metric records when a
- * free-text query is present. When there is no query (e.g. browsing by
- * topic), FMs are kept so they can surface at the top of topic pages.
- */
-export function formatFeaturedMetricFacetFilter(
-    query: string
-): SearchFacetFilters {
-    return query.trim() ? ["isFM:false"] : []
-}
-
-export function formatCountryFacetFilters(
-    countries: Set<string>,
-    requireAllCountries: boolean
-) {
-    const facetFilters: SearchFacetFilters = []
-    if (requireAllCountries) {
-        facetFilters.push(
-            ...formatConjunctiveFacetFilters(countries, "availableEntities")
-        )
-    } else {
-        facetFilters.push(
-            ...formatDisjunctiveFacetFilters(countries, "availableEntities")
-        )
-    }
-    // Don't show income group-specific FMs if no countries are selected
-    if (!countries.size) {
-        facetFilters.push("isIncomeGroupSpecificFM:false")
-    }
-    return facetFilters
-}
-
-export const formatTopicFacetFilters = (
-    topics: Set<string>
-): SearchFacetFilters => {
-    return formatDisjunctiveFacetFilters(topics, "tags")
-}
-
 export function serializeSet(set: Set<string>) {
     return set.size ? [...set].join("~") : undefined
 }
 
 export function deserializeSet(str: string | null): Set<string> {
     return str ? new Set(str.split("~")) : new Set()
-}
-
-export function getFilterNamesOfType(
-    filters: Filter[],
-    type: FilterType
-): Set<string> {
-    return new Set(
-        filters
-            .filter((filter) => filter.type === type)
-            .map((filter) => filter.name)
-    )
 }
 
 export const getFilterIcon = (filter: Filter) => {
